[Refactor]: ProductImageService 생성
