+++ src/main/java/com/backend/domain/product/service/ProductService.java
@@ -1,15 +1,13 @@
 package com.backend.domain.product.service;
 
 import com.backend.domain.member.entity.Member;
+import com.backend.domain.product.dto.ProductSearchDto;
 import com.backend.domain.product.dto.request.ProductCreateRequest;
 import com.backend.domain.product.dto.request.ProductModifyRequest;
-import com.backend.domain.product.dto.ProductSearchDto;
 import com.backend.domain.product.entity.Product;
-import com.backend.domain.product.entity.ProductImage;
 import com.backend.domain.product.enums.*;
-import com.backend.domain.product.repository.ProductImageRepository;
-import com.backend.domain.product.repository.ProductRepository;
 import com.backend.domain.product.exception.ProductException;
+import com.backend.domain.product.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -21,32 +19,29 @@ import org.springframework.web.multipart.MultipartFile;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
 public class ProductService {
+    private final ProductImageService productImageService;
     private final ProductRepository productRepository;
-    private final ProductImageRepository productImageRepository;
-    private final FileService fileService;
 
     // ======================================= create methods ======================================= //
     @Transactional
-    public Product create(Member actor, ProductCreateRequest request, List<MultipartFile> images) {
-        // 0. 유효성 검증 (location, images)
+    public Product createProduct(Member actor, ProductCreateRequest request, List<MultipartFile> images) {
+        // location 검증
         validateLocation(request.location(), request.deliveryMethod());
-        validateImagesForCreate(images);
 
-        // 1. Product 생성 및 저장
-        Product savedProduct = createProduct(actor, request);
+        // Product 생성 및 저장
+        Product savedProduct = saveProduct(actor, request);
 
-        // 2. 이미지 업로드 및 저장
-        createProductImages(savedProduct, images);
+        // 이미지 검증 및 저장
+        productImageService.validateAndCreateImages(savedProduct, images);
 
         return savedProduct;
     }
 
-    public Product createProduct(Member actor, ProductCreateRequest request) {
+    public Product saveProduct(Member actor, ProductCreateRequest request) {
         Product product = new Product(
                 request.name(),
                 request.description(),
@@ -61,24 +56,7 @@ public class ProductService {
         return productRepository.save(product);
     }
 
-    private void createProductImages(Product savedProduct, List<MultipartFile> images) {
-        for (MultipartFile image : images) {
-            String imageUrl = fileService.uploadFile(image, "products/" + savedProduct.getId());
-            createProductImage(savedProduct, imageUrl);
-        }
-    }
-
-    public void createProductImage(Product savedProduct, String imageUrl) {
-        ProductImage productImage = new ProductImage(imageUrl, savedProduct);
-        productImageRepository.save(productImage);
-        savedProduct.addProductImage(productImage);
-    }
-
     // ======================================= find/get methods ======================================= //
-    public Optional<Product> findLatest() {
-        return productRepository.findFirstByOrderByIdDesc();
-    }
-
     @Transactional(readOnly = true)
     public Page<Product> findBySearchPaged(
             int page, int size, ProductSearchSortType sort, ProductSearchDto search
@@ -95,8 +73,10 @@ public class ProductService {
         return productRepository.findByMemberPaged(pageable, actor.getId(), SaleStatus.fromSaleStatus(status));
     }
 
-    public long count() {
-        return productRepository.count();
+    private Pageable getPageable(int page, int size, ProductSearchSortType sort) {
+        page = (page > 0) ? page : 1;
+        size = (size > 0 && size <= 100) ? size : 20;
+        return PageRequest.of(page - 1, size, sort.toSort());
     }
 
     public Optional<Product> findById(Long productId) {
@@ -107,50 +87,38 @@ public class ProductService {
         return findById(productId).orElseThrow(ProductException::notFound);
     }
 
-    private Pageable getPageable(int page, int size, ProductSearchSortType sort) {
-        page = (page > 0) ? page : 1;
-        size = (size > 0 && size <= 100) ? size : 20;
-        return PageRequest.of(page - 1, size, sort.toSort());
+    public Optional<Product> findLatest() {
+        return productRepository.findFirstByOrderByIdDesc();
+    }
+
+    public long count() {
+        return productRepository.count();
     }
 
     // ======================================= modify/delete methods ======================================= //
     @Transactional
     public Product modifyProduct(Product product, ProductModifyRequest request, List<MultipartFile> images, List<Long> deleteImageIds) {
+        // 수정 요청 검증
         ProductModifyRequest validatedRequest = validateModifyRequest(product, request);
-        validateImagesForModify(product, images);
 
+        // Product 수정
         product.modify(validatedRequest);
 
-        if (images != null && !images.isEmpty()) {
-            createProductImages(product, images);
-            productImageRepository.flush();
-        }
-
-        if (deleteImageIds != null) {
-            for (Long deleteImageId : deleteImageIds) {
-                ProductImage productImage = productImageRepository.findById(deleteImageId).orElseThrow(ProductException::imageNotFound);
-                if (!productImage.getProduct().getId().equals(product.getId())) throw ProductException.imageNotBelongToProduct();
-
-                fileService.deleteFile(productImage.getImageUrl());
-                product.deleteProductImage(productImage);
-                productImageRepository.delete(productImage);
-            }
-
-            if (product.getProductImages().isEmpty()) throw ProductException.imageRequired();
-        }
+        // 이미지 검증 및 수정
+        productImageService.validateAndModifyImages(product, images, deleteImageIds);
 
         return product;
     }
 
+    @Transactional
     public void deleteProduct(Product product) {
+        // 삭제 권한 검증
         if (product.getStartTime().isBefore(LocalDateTime.now())) {
             throw ProductException.auctionDeleteForbidden();
         }
 
-        for (ProductImage pi : product.getProductImages()) {
-            fileService.deleteFile(pi.getImageUrl());
-        }
-
+        // 이미지 및 Product 삭제
+        productImageService.deleteAllProductImages(product);
         productRepository.delete(product);
     }
 
@@ -159,55 +127,6 @@ public class ProductService {
         if (deliveryMethod == DeliveryMethod.TRADE || deliveryMethod == DeliveryMethod.BOTH) {
             if (location == null || location.isBlank()) {
                 throw ProductException.locationRequired();
-            }
-        }
-    }
-
-    private void validateImagesForCreate(List<MultipartFile> images) {
-        if (images == null || images.isEmpty()) {
-            throw ProductException.imageRequired();
-        }
-
-        if (images.size() > 5) {
-            throw ProductException.imageMaxCountExceeded();
-        }
-
-        validateImages(images);
-    }
-
-    private void validateImagesForModify(Product product, List<MultipartFile> images) {
-        if (images == null || images.isEmpty()) {
-            return;
-        }
-        if (images.size() + product.getProductImages().size() > 5) {
-            throw ProductException.imageMaxCountExceeded();
-        }
-
-        validateImages(images);
-    }
-
-    private void validateImages(List<MultipartFile> images) {
-        Set<String> allowedExtensions = Set.of(".jpg", ".jpeg", ".png", ".gif", ".webp");
-
-        for (MultipartFile image : images) {
-            if (image.isEmpty()) {
-                throw ProductException.emptyFile();
-            }
-
-            // 파일 크기 검증
-            if (image.getSize() > 5 * 1024 * 1024) {
-                throw ProductException.fileTooLarge();
-            }
-
-            // 파일 확장자 검증
-            String filename = image.getOriginalFilename();
-            if (filename == null || !filename.contains(".")) {
-                throw ProductException.invalidFileName();
-            }
-
-            String extension = filename.substring(filename.lastIndexOf(".")).toLowerCase();
-            if (!allowedExtensions.contains(extension)) {
-                throw ProductException.unsupportedFileType();
             }
         }
     }
